refactor: move InputText and InputSelect from ProfileModal to UI folder
