+++ src/components/UI/InputText/InputText.tsx
@@ -1,15 +1,21 @@
 import React, { FC } from 'react';
 import { TextField } from '@mui/material';
-import { IInputTextProps } from './InputText.types';
+import { IInputTextProps } from './InputText.type';
 
-export const InputText: FC<IInputTextProps> = ({ name, register, registerName, ...props }) => {
+export const InputText: FC<IInputTextProps> = ({
+  name,
+  register,
+  registerName,
+  size = 'small',
+  ...props
+}) => {
   return (
     <TextField
       {...props}
       inputProps={register(registerName)}
       name={name}
       fullWidth
-      size="small"
+      size={size}
       margin="normal"
       label={name}
     />
