--- src/pages/EmployeesProfilePage/ProfileModal/InputText/InputText.types.ts
@@ -1,10 +1,0 @@
-import { UseFormRegister } from 'react-hook-form';
-import { IProfileFormInput, TInputTextType } from '../ProfileModal.types';
-
-export interface IInputTextProps {
-  name: string;
-  registerName: TInputTextType;
-  register: UseFormRegister<IProfileFormInput>;
-  error?: boolean;
-  helperText?: string;
-}
