+++ src/types/errorTypes.ts
@@ -1,0 +1,1 @@
+export type TError = Record<string, string>;
