+++ src/types/formTypes.ts
@@ -1,0 +1,3 @@
+import { FieldValues } from 'react-hook-form';
+
+export type TFormSubmit = (data: FieldValues) => Promise<void>;
